requirements.txt
dash ~= 1.9
pillow
